requirements.txt
# Web Framework
Flask==3.0.3

# Web Scraping
selenium==4.22.0
beautifulsoup4==4.12.3
requests==2.32.3

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.31

# Environment Variables
python-dotenv==1.0.1

# Data Manipulation
pandas==2.2.2

# Machine Learning
transformers==4.42.1
torch==2.3.1 # Or tensorflow, depending on preference and system.
scikit-learn==1.5.0

# Dashboard
streamlit==1.36.0
wordcloud==1.9.3

# Logging
loguru==0.7.2
